+++ lib/src/cipherlib_base.dart
@@ -3,4 +3,6 @@
 
 export 'chacha20.dart';
 export 'chacha20_poly1305.dart';
+export 'salsa20.dart';
+export 'salsa20_poly1305.dart';
 export 'xor.dart';