+++ lib/src/salsa20.dart
@@ -3,48 +3,42 @@
 
 import 'dart:typed_data';
 
-import 'package:cipherlib/src/algorithms/chacha20.dart';
+import 'package:cipherlib/src/algorithms/salsa20.dart';
 
-export 'algorithms/chacha20.dart' show ChaCha20;
+export 'algorithms/salsa20.dart' show Salsa20;
 
-/// Apply [ChaCha20] cipher with the follwing parameters:
+/// Apply [Salsa20] cipher with the follwing parameters:
 ///
 /// - Arbitrary length plaintext [message] to transform.
 /// - A 16 or 32-bytes long [key].
 /// - (Optional) A 16-bytes long [nonce]. Default: 0
-/// - (Optional) The initial block number as [blockCount]. Default: 1.
 ///
 /// Both the encryption and decryption can be done using this same method.
 @pragma('vm:prefer-inline')
-Uint8List chacha20(
+Uint8List salsa20(
   List<int> message,
   List<int> key, [
   List<int>? nonce,
-  int blockCount = 1,
 ]) =>
-    ChaCha20(key).convert(
+    Salsa20(key).convert(
       message,
       nonce: nonce,
-      blockCount: blockCount,
     );
 
-/// Apply [ChaCha20] cipher with the follwing parameters:
+/// Apply [Salsa20] cipher with the follwing parameters:
 ///
 /// - Plaintext message [stream] to transform.
 /// - A 16 or 32-bytes long [key].
 /// - (Optional) A 16-bytes long [nonce]. Default: 0
-/// - (Optional) The initial block number as [blockCount]. Default: 1.
 ///
 /// Both the encryption and decryption can be done using this same method.
 @pragma('vm:prefer-inline')
-Stream<int> chacha20Pipe(
+Stream<int> salsa20Pipe(
   Stream<int> stream,
   List<int> key, [
   List<int>? nonce,
-  int blockCount = 1,
 ]) =>
-    ChaCha20(key).pipe(
+    Salsa20(key).pipe(
       stream,
       nonce: nonce,
-      blockId: blockCount,
     );